requirements.txt
aiofiles==0.8.0
requests==2.27.1
paramiko==2.11.0
python-telegram-bot
werkzeug==2.0.3
pytz
aiohttp
APScheduler
websockets
asyncssh
hypercorn
asgiref
quart
cryptography
